+++ server/farm/src/main/java/com/bzzzzz/farm/cart/controller/CartController.java
@@ -1,17 +1,19 @@
 package com.bzzzzz.farm.cart.controller;
 
+import com.bzzzzz.farm.cart.dto.CartProductPatchDto;
+import com.bzzzzz.farm.cart.dto.CartProductPostDto;
 import com.bzzzzz.farm.cart.entiy.Cart;
 import com.bzzzzz.farm.cart.mapper.CartMapper;
 import com.bzzzzz.farm.cart.service.CartService;
+import com.bzzzzz.farm.dto.IdRequestDto;
+import com.bzzzzz.farm.dto.SingleResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 
 @RestController
@@ -29,5 +31,32 @@ public class CartController {
         Cart cart = cartService.findCart(cartId);
 
         return new ResponseEntity(cartMapper.cartToCartResponseDto(cart), HttpStatus.OK);
+    }
+
+    @PostMapping("/products")
+    public ResponseEntity postCartProduct(@Valid @RequestBody CartProductPostDto cartProductPostDto) {
+        //Todo: 로그인 관련 기능 들어오면 로그인했는지 확인하는 로직 필요
+
+        cartService.createCartProduct(cartMapper.cartProductPostDtoToCartProduct(cartProductPostDto));
+
+        return new ResponseEntity<>(new SingleResponseDto(cartProductPostDto.getCartId()), HttpStatus.CREATED);
+    }
+
+    @PatchMapping("/products")
+    public ResponseEntity patchCartProduct(@Valid @RequestBody CartProductPatchDto cartProductPatchDto) {
+        //Todo: 로그인 관련 기능 들어오면 로그인했는지 확인하는 로직 필요
+
+        cartService.updateCartProduct(cartProductPatchDto.getCartProductId(), cartProductPatchDto.getQuantity());
+
+        return new ResponseEntity<>(new SingleResponseDto(cartProductPatchDto.getCartId()), HttpStatus.OK);
+    }
+
+    @DeleteMapping("/products")
+    public ResponseEntity deleteCartProduct(@Valid @RequestBody IdRequestDto idRequestDto) {
+        //Todo: 로그인 관련 기능 들어오면 로그인했는지 확인하는 로직 필요
+
+        cartService.deleteCartProduct(idRequestDto.getId());
+
+        return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
 }
