+++ server/farm/src/main/java/com/bzzzzz/farm/cart/service/CartService.java
@@ -1,6 +1,8 @@
 package com.bzzzzz.farm.cart.service;
 
 import com.bzzzzz.farm.cart.entiy.Cart;
+import com.bzzzzz.farm.cart.entiy.CartProduct;
+import com.bzzzzz.farm.cart.repository.CartProductRepository;
 import com.bzzzzz.farm.cart.repository.CartRepository;
 import com.bzzzzz.farm.exception.BusinessLogicException;
 import com.bzzzzz.farm.exception.ExceptionCode;
@@ -15,16 +17,52 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class CartService {
     private final CartRepository cartRepository;
+    private final CartProductRepository cartProductRepository;
 
+    @Transactional(readOnly = true)
     public Cart findCart(long cartId) {
         return findVerifiedCart(cartId);
+    }
+
+    public void createCartProduct(CartProduct cartProduct) {
+        Optional<CartProduct> optionalCartProduct = cartProductRepository.findByCartAndProductOption(cartProduct.getCart(), cartProduct.getProductOption());
+
+        if (optionalCartProduct.isPresent()) {
+            updateCartProduct(optionalCartProduct.get().getCartProductId(), cartProduct.getQuantity());
+        } else {
+            cartProductRepository.save(cartProduct);
+        }
+    }
+
+    public void updateCartProduct(long cartProductId, Integer quantity) {
+        CartProduct findCartProduct = findVerifiedCartProduct(cartProductId);
+
+        // Optional.ofNullable(quantity).ifPresent(data -> findCartProduct.calculateQuantity(data));
+        findCartProduct.calculateQuantity(quantity); // 장바구니 수정항목이 추가 되면 옵셔널로 변경하시오
+
+        if (findCartProduct.getQuantity() <= 0) {
+            deleteCartProduct(findCartProduct.getCartProductId());
+        }
+    }
+
+    public void deleteCartProduct(long cartProductId) {
+        CartProduct cartProduct = findVerifiedCartProduct(cartProductId);
+
+        cartProductRepository.delete(cartProduct);
     }
 
     /**
      * 서브 메서드
      */
+    @Transactional(readOnly = true)
     private Cart findVerifiedCart(long cartId) {
         Optional<Cart> optionalCart = cartRepository.findById(cartId);
         return optionalCart.orElseThrow(() -> new BusinessLogicException(ExceptionCode.CART_NOT_FOUND));
+    }
+
+    @Transactional(readOnly = true)
+    private CartProduct findVerifiedCartProduct(long cartProductId) {
+        Optional<CartProduct> optionalCartProduct = cartProductRepository.findById(cartProductId);
+        return optionalCartProduct.orElseThrow(() -> new BusinessLogicException(ExceptionCode.CART_PRODUCT_NOT_FOUND));
     }
 }
